WIP: minor updates in ODF.
